Transfer graphing from BioMass to Text2Model
